refactor: ors 관련 파일을 Infra 폴더로 분리
