+++ src/main/java/com/ridingmate/api_server/domain/route/facade/RouteFacade.java
@@ -6,9 +6,9 @@ import com.ridingmate.api_server.domain.route.dto.response.CreateRouteResponse;
 import com.ridingmate.api_server.domain.route.dto.response.RouteSegmentResponse;
 import com.ridingmate.api_server.domain.route.entity.Route;
 import com.ridingmate.api_server.domain.route.service.RouteService;
-import com.ridingmate.api_server.global.client.OrsClient;
-import com.ridingmate.api_server.global.client.OrsMapper;
-import com.ridingmate.api_server.global.client.dto.response.OrsRouteResponse;
+import com.ridingmate.api_server.infra.ors.OrsClient;
+import com.ridingmate.api_server.infra.ors.OrsMapper;
+import com.ridingmate.api_server.infra.ors.dto.response.OrsRouteResponse;
 import com.ridingmate.api_server.global.util.GeometryUtil;
 import lombok.RequiredArgsConstructor;
 import org.locationtech.jts.geom.LineString;
